🔨 Fix ShowCreateDialog.vue this.datetime is not ""

See merge request s05-webmobile1-sub3/S05P13A405!156

+++ backend/src/main/java/com/ssafy/spotlive/api/service/ShowInfoServiceImpl.java
@@ -89,7 +89,8 @@ public class ShowInfoServiceImpl implements ShowInfoService {
         if (posterImage != null && !posterImage.isEmpty()) {
             String nextPosterImageUrl = null;
             String currentPosterUrl = showInfo.getPosterUrl();
-            fileUploadService.delete(currentPosterUrl);
+            if (currentPosterUrl != null)
+                fileUploadService.delete(currentPosterUrl);
             try {
                 nextPosterImageUrl = fileUploadService.upload(posterImage);
                 showInfo.setPosterUrl(nextPosterImageUrl);
